scripts/requirements.txt
# Requirements for synthetic log generator scripts
# Install with: pip install -r requirements.txt

portkey-ai>=1.0.0
python-dotenv>=1.0.0
